+++ app/src/main/java/com/abhi/noteIt/Database/NoteDatabase.java
@@ -1,4 +1,4 @@
-package com.abhi.noteIt;
+package com.abhi.noteIt.Database;
 
 import android.content.Context;
 import android.os.AsyncTask;
@@ -8,6 +8,8 @@ import androidx.room.Database;
 import androidx.room.Room;
 import androidx.room.RoomDatabase;
 import androidx.sqlite.db.SupportSQLiteDatabase;
+
+import com.abhi.noteIt.Model.Note;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
